Dialogue and Player HealthBar
dialogue system. Consists of three parts; Dialoguetrigger, dialogue, DialogueManager
    The dialogue box would stay inactive until called upon by the dialoguemanager.
    Dialogue trigger would be what the objects or instances instantializing the dialogue would have. This trigger would then call the dialogue manager.
    dialogue itself is only a data structure. Has name and the paragraph/sentence array.

    Right now there is only a button to test and it works. Set the button to active before starting scene to test it.
   Added it into Alice Scene 1.

    also has comments but lemme know if you have any questions or concerns

    but right now, next step is to incoporate it to be used with events rather than a button push

Pushed updated version of the player health bar as well. Added it to Alice Test Scene 1 and made it more uniform with the boss. (Player health is above their head)

Otherwise let me know if we prefer it in the corner and I can go back and change it no problem