How To Play + Pause Menu
How To Play button in Title Menu added with basic controls.

A menu button has been added on the top right of the screen on the goat boss level. It uses the same 2 page sprite, with buttons allowing the user to restart the level, quit the game, or go back. Clicking the menu button will pause the game and show the basic controls on the right page, with interactive buttons on the left page.