player can now stun the enemy without taking damage and added barriers
the barriers have a max and min that represent the x of the Vector3. It is used to reposition at the start Barriers on start and end sides. The script createbarrier was modified so that at the start based on the inputted max and min float it will change the position.x